+++ .gitignore
@@ -55,3 +55,4 @@ freeline/
 freeline_project_description.json
 /dist/
 /nbproject/private/
+/target/
