--- nbproject/private/private.properties
@@ -1,7 +1,0 @@
-application.args=19-01-2018_17:11:11  24-01-2018_13:01:01
-compile.on.save=true
-do.depend=false
-do.jar=true
-javac.debug=true
-javadoc.preview=true
-user.properties.file=C:\\Users\\dev1\\AppData\\Roaming\\NetBeans\\8.2\\build.properties
