+++ src/main/java/ru/osk/sd/Weekends.java
@@ -1,0 +1,31 @@
+/*
+ * To change this license header, choose License Headers in Project Properties.
+ * To change this template file, choose Tools | Templates
+ * and open the template in the editor.
+ */
+package ru.osk.sd;
+
+import java.util.Calendar;
+
+/**
+ *
+ * @author dev1
+ */
+public enum Weekends {
+    SUNDAY(Calendar.SUNDAY),
+    MONDAY(Calendar.MONDAY),
+    TUESDAY(Calendar.TUESDAY),
+    WEDNESDAY(Calendar.WEDNESDAY),
+    THURSDAY(Calendar.THURSDAY),
+    FRIDAY(Calendar.FRIDAY),
+    SATURDAY(Calendar.SATURDAY);
+    
+    private final int id;
+    Weekends(int id) {
+        this.id = id; 
+    }
+    
+    public int getValue() { 
+        return id; 
+    }
+}
